refactor: Organise code to separate subfolders
