+++ serverless/database-backup/src/backups/index.ts
@@ -1,0 +1,1 @@
+export { default as Pgdump } from './pgdump'
