+++ serverless/database-backup/src/scripts/decrypt.ts
